+++ tls-utils/Cargo.toml
@@ -2,6 +2,7 @@
 name = "solana-tls-utils"
 description = "Solana TLS utilities"
 documentation = "https://docs.rs/solana-tls-utils"
+publish = false
 
 version.workspace = true
 authors.workspace = true
@@ -11,14 +12,13 @@ license.workspace = true
 edition.workspace = true
 
 [dependencies]
-solana-sdk = { workspace = true }
 
 rustls = { workspace = true, features=["ring"] }
-x509-parser = { workspace = true }
 solana-keypair = { workspace = true }
-solana-signer = { workspace = true }
 solana-pubkey = { workspace = true }
-
+solana-sdk = { workspace = true }
+solana-signer = { workspace = true }
+x509-parser = { workspace = true }
 
 [lints]
 workspace = true
