+++ tls-utils/README
@@ -1,1 +1,1 @@
-A collection of utility functions and structures needed to bridge the conceptual gap between conventional TLS security model in protocols like QUIC and what Solana does
+A collection of utility functions and structures needed to bridge the conceptual gap between conventional TLS security models in protocols like QUIC and Agave use cases
