+++ tls-utils/src/lib.rs
@@ -1,5 +1,5 @@
 //! Collection of TLS related code fragments that end up popping up everywhere where quic is used.
-//! Aggregated here to avoid bugs due to conflicting implementations of the same functionality
+//! Aggregated here to avoid bugs due to conflicting implementations of the same functionality.
 
 mod tls_certificates;
 pub use tls_certificates::*;
@@ -9,3 +9,6 @@ pub use quic_client_certificate::*;
 
 mod skip_server_verification;
 pub use skip_server_verification::SkipServerVerification;
+
+mod skip_client_verification;
+pub use skip_client_verification::SkipClientVerification;
